+++ ts/libs/core/src/lib/slo/public/common/slo-output.ts
@@ -1,4 +1,4 @@
-import { SloMapping } from '../../../model';
+import { ElasticityStrategyKind, SloMapping } from '../../../model';
 
 /**
  * Describes the output of `ServiceLevelObjective.execute()`.
@@ -18,6 +18,20 @@ export interface SloOutput<T> {
      * The `ElasticityStrategySpec.sloOutputParams` that should be submitted to the cluster.
      */
     elasticityStrategyParams: T;
+
+    /**
+     * (optional) The `ElasticityStrategyKind` that identifies the ElasticityStrategy that should be triggered.
+     *
+     * This only needs to be set, if an ElasticityStrategy other than the default strategy specified in the `SloMapping` should be triggered.
+     */
+    elasticityStrategy?: ElasticityStrategyKind<T>;
+
+    /**
+     * (optional) Static configuration for the `ElasticityStrategy`.
+     *
+     * This only needs to be set, if other values than the default ones specified in the `SloMapping` should be used.
+     */
+    staticElasticityStrategyConfig?: any;
 
     // We could use this to delete a strategy in the future.
     // deleteStrategy?: boolean;